[DISCO-3036] refactor: r2d2 job to use domain mapping file
